test: use node-core-test instead of vitest
