+++ tests/node/index.test.ts
@@ -1,16 +1,22 @@
 /* eslint @typescript-eslint/ban-ts-comment: off */
-import assert from 'node:assert/strict'
-import { createHmac, webcrypto } from 'node:crypto'
-import { describe, it } from 'node:test'
+import { Crypto } from '@peculiar/webcrypto'
+import assert from 'assert'
+import crypto from 'crypto'
 import { satisfies } from 'semver'
+import test from 'test'
 import { tests } from '../index.js'
 
 tests({
   // @ts-expect-error
-  crypto: satisfies(process.version, '>=19.0.0') ? crypto : webcrypto,
-  createHmac,
-  describe,
-  it,
-  deepEqual: assert.deepEqual,
+  // eslint-disable-next-line no-nested-ternary
+  crypto: satisfies(process.version, '>=19.0.0')
+    ? globalThis.crypto
+    : satisfies(process.version, '>=16.0.0')
+    ? crypto.webcrypto
+    : new Crypto(),
+  createHmac: crypto.createHmac,
+  describe: test.describe,
+  it: test.it,
+  deepEqual: assert.deepStrictEqual,
   rejects: assert.rejects,
 })
