--- tests/vite-node/index.test.ts
@@ -1,25 +1,0 @@
-/* eslint @typescript-eslint/ban-ts-comment: off */
-import { Crypto } from '@peculiar/webcrypto'
-import crypto from 'node:crypto'
-import { satisfies } from 'semver'
-import { describe, expect, it } from 'vitest'
-import { tests } from '../index.js'
-
-tests({
-  // @ts-expect-error
-  // eslint-disable-next-line no-nested-ternary
-  crypto: satisfies(process.version, '>=19.0.0')
-    ? globalThis.crypto
-    : satisfies(process.version, '>=16.0.0')
-    ? crypto.webcrypto
-    : new Crypto(),
-  createHmac: crypto.createHmac,
-  describe,
-  it,
-  deepEqual(actual, expected) {
-    expect(actual).toStrictEqual(expected)
-  },
-  async rejects(fn, re) {
-    return expect(fn).rejects.toThrow(re)
-  },
-})
